+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/impl/parse/EnsemblParser.java
@@ -71,13 +71,13 @@ public class EnsemblParser implements TranscriptParser {
 		}
 
 		// Parse the GFF file and feed the resulting Feature objects into a TranscriptModelBuilder.
-		FeatureProcessor fp = new FeatureProcessor(gffParser.getGffVersion(), refDict);
+		FeatureProcessor fp = new FeatureProcessor(gffParser.getGffVersion(), true, refDict);
 		gffParser.parse(fp);
 		// Build ArrayList of TranscriptModelBuilder objects from feature list.
 		ArrayList<TranscriptModelBuilder> builders;
 		try {
 			LOGGER.info("Building transcript models...");
-			TranscriptInfoFactory tif = new TranscriptInfoFactory(gffParser.getGffVersion(), refDict);
+			TranscriptInfoFactory tif = new TranscriptInfoFactory(true, gffParser.getGffVersion(), refDict);
 			builders = tif.buildTranscripts(fp.getGenes());
 			TranscriptSupportLevelsSetterFromLengths.run(builders);
 		} catch (InvalidAttributeException e) {